create search blueprint
