+++ app/search/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+search_bp = Blueprint('search', __name__)
+
+from app.search import routes
